Add Bintray/JCenter build artifact deployment
[Note]
This configuration was not fully tested yet, since the link to
JCenter still has to be approved. The pom.xml file is already
correctly created and uploading to Bintray works with correct
authentication information. A guide on how to publish the
library to Bintray has been added, but might be changed in
the future, if the published artifact doesn't live up to
our expectations.